+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 
+### Added
+
+- Support for the `LendingProtocolV1_1` amendment.
+
 ## [[5.1.0]]
 
 ### Added